Touchscreen tests improvements
